requirements.txt
flask>=3.0.3
flask-cors>=5.0.1
gunicorn>=23.0.0
requests>=2.32.3
python-chess>=1.999
pytest>=8.3.5
